Update Dockerfile
Upgrading the PHP image version to 8.0 because PECL threw an error on Linux Kernel 5.13. Unsure if this is an update on the PECL side or because of the different kernel, though I suspect the latter since even a week ago this didn't exist until testing on this new Kernel version.